requirements.txt
Flask
requests
google-genai
flask-cors
